Add an empty line after content before the link list.
The transformer was generating content and the link list without adding
an empty line after the content. The integration tests were passing due
the fact that the last content element in input was a table and it
enforces an empty line. New implementation chooses lesser of the evils
and creates an extra insignificant new line in such cases.